bled: flash: consider timer max value on calculation
Consider timer max value on tick calulcations. This is actually only
needed when passing the compare tick value to to timer, cause it will
thrown an assert when the value is greater than the max. value. In
all other cases, it's just an overflow.

+++ src/bled.c
@@ -356,11 +356,11 @@ static int bled_hw_ctrl_cmd_flash(bled_led_t bled, bled_flash_cmd_arg_t* cmd_arg
 
 	CORE_ATOMIC_SECTION(
 		tmr_cnt_tick = TIMER_CounterGet(bled_p->hw_tmr.base);
-		bled_p->hw_tmr.ldma_buf[0] = tmr_cnt_tick
+		bled_p->hw_tmr.ldma_buf[0] = (uint16_t)(tmr_cnt_tick
 			+ tmr_delay_ticks
-			+ tmr_flash_ticks;
-		TIMER_CompareSet(bled_p->hw_tmr.base, bled_p->hw_tmr.ch, tmr_cnt_tick
-			+ tmr_delay_ticks);
+			+ tmr_flash_ticks);
+		TIMER_CompareSet(bled_p->hw_tmr.base, bled_p->hw_tmr.ch, (uint16_t)
+			(tmr_cnt_tick + tmr_delay_ticks));
 
 		TIMER_RouteCCSet(bled_p->hw_tmr.base, bled_p->hw_tmr.ch,
 			bled_p->hw_tmr.route);
